organize style queries
