+++ components/Header/DesktopHeader.tsx
@@ -70,7 +70,7 @@ const Wrapper = styled.div`
   justify-content: space-between;
   align-items: center;
   height: 40px;
-  max-width: var(--max-section-width);
+  max-width: var(--desktop-page-width);
   margin: 0 auto;
   z-index: 100;
   border-radius: 12px;
