+++ constant/style/breakpoints.ts
@@ -1,0 +1,3 @@
+export const mobile = 640;
+export const tablet = 1024;
+export const laptop = 1300;
