+++ constant/style/containers.ts
@@ -1,0 +1,1 @@
+export const desktopPageWidth = 1140;
