+++ constant/style/queries.ts
@@ -1,9 +1,6 @@
-export const maxSectionWidth = 1140;
-
-export const mobile = 640;
-export const tablet = 1024;
-export const laptop = 1300;
+import { laptop, mobile, tablet } from "./breakpoints";
 
 export const mobileAndUnder = `(max-width: ${mobile}px)`;
 export const tabletAndUnder = `(max-width: ${tablet}px)`;
 export const laptopAndUnder = `(max-width: ${laptop}px)`;
+export const onlyTablet = `(min-width: ${mobile + 1}px) and (max-width: ${tablet}px)`;
